feat(roundsLogic): refactor game logic for cardsRemain and RoundsWon
