+++ app/gameLogic/index.js
@@ -4,3 +4,4 @@ export * from './gameCreationFunctions'
 export * from './cardPickingFunctions'
 export * from './roleSpecificationFunctions'
 export * from './roundManagementFunctions'
+
