+++ app/gameLogic/roundManagementFunctions.js
@@ -1,32 +1,20 @@
-import firebase from 'APP/fire'
 
-// When Cards Remaining for either team hits 0 then call RoundsWonIncrement Logic
-
-// function roundsWonIncrement(currentWinningTeam, gameId) {
-//   const dataRef = firebase.database().ref()
-//   const gameInstance = dataRef.child('gameInstances').child(gameId).child('currentGameStatus').child('RoundsWonByTeams')
-
-//   if (currentWinningTeam === 'redTeam') {
-
-//   }
-//   gameInstance.once('value').then
-
-//   gameInstance.update({
-//     blueTeamNumRoundsWon: 0,
-//     redTeamNumRoundsWon:
-//   })
-
-// }
-
-const getCurrentRoundsWonScore = function(gameId) {
-  const dataRef = firebase.database().ref()
-  const gameInstance = dataRef.child('gameInstances').child(gameId).child('currentGameStatus').child('RoundsWonByTeams')
-  return gameInstance.once('value').then(function(snap) {
-    const roundsWon = snap.val()
-    const redTeamRoundsWon = roundsWon.redTeamNumRoundsWon
-    const blueTeamRoundsWon = roundsWon.blueTeamNumRoundsWon
-    return [redTeamRoundsWon, blueTeamRoundsWon]
-  })
+export const updateRoundsWon = function(blueCardsLeft, redCardsLeft, blueRoundsWon, redRoundsWon) {
+  let currPhaseOfGame = {}
+  if (blueCardsLeft === 0) {
+    currPhaseOfGame = {
+      RoundsWonByTeams: {
+        blueTeamNumRoundsWon: blueRoundsWon + 1,
+        redTeamNumRoundsWon: redRoundsWon
+      }
+    }
+  } else if (redCardsLeft === 0) {
+    currPhaseOfGame = {
+      RoundsWonByTeams: {
+        blueTeamNumRoundsWon: blueRoundsWon,
+        redTeamNumRoundsWon: redRoundsWon + 1
+      }
+    }
+  }
+  return currPhaseOfGame
 }
-
-export default getCurrentRoundsWonScore
